reorganize files & fix dependency cycle
