+++ app/javascript/react/components/FileUpload/FailedUrlList/FailedUrlList.jsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 
-import Url from './Url/Url';
+import Url from './Url';
 
 const failed_url_list = (props) => (
   <div>
