+++ app/javascript/react/components/FileUpload/FileList/File.jsx
@@ -1,10 +1,16 @@
 import React from 'react';
 
-import {TabularCheckStatus} from '../../../../containers/FileUpload/UploadFiles';
-
-import ellipsize from '../../../../../lib/string_patch';
+import ellipsize from '../../../../lib/string_patch';
 
 import classes from './File.module.css';
+
+export const TabularCheckStatus = {
+  checking: 'Checking...',
+  issues: 'View Issues',
+  noissues: 'Passed',
+  na: 'Too Large For Validation',
+  error: "Couldn't Read Tabular Data",
+};
 
 const capitalize = (str) => str.charAt(0).toUpperCase() + str.slice(1);
 
@@ -21,7 +27,7 @@ const statusCss = (status) => {
   }
 };
 
-const file = (props) => {
+function File(props) {
   let tabularInfo;
 
   if (props.removingIndex !== props.index) {
@@ -108,6 +114,6 @@ const file = (props) => {
         ) }
     </tr>
   );
-};
+}
 
-export default file;
+export default File;
