+++ app/javascript/react/components/FileUpload/FileList/File.module.css
