+++ app/javascript/react/components/FileUpload/ModalUrl.jsx
@@ -2,7 +2,7 @@
 // This may be erroring since the label is hard coded, I guess bad practice if used more than once on page, but this isn't.
 
 import React from 'react';
-import ValidateFiles from '../ValidateFiles/ValidateFiles';
+import ValidateFiles from './ValidateFiles';
 
 const modal_url = (props) => (
   <div>
