+++ app/javascript/react/components/FileUpload/UploadType.jsx
