+++ app/javascript/react/containers/index.js
@@ -1,0 +1,1 @@
+export {default as UploadFiles } from './FileUpload/UploadFiles'
